+++ CHANGELOG.md
@@ -17,8 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-### Security
+* `NormalizedString` accepting empty usernames and passwords.
 
+### Security
 
 ## [0.4.1] - 2022-05-09